% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{add_colnames}
\alias{add_colnames}
\title{Add column names to matrices in the list x}
\usage{
add_colnames(x, colnames)
}
\arguments{
\item{x}{A list of matrices and other elements (ignored)}

\item{colnames}{A list of column names to set for all matrices}
}
\description{
Add column names to matrices in the list x
}
