% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_ensembl}
\alias{annotate_ensembl}
\title{Annotate object indexed by ensembl}
\usage{
annotate_ensembl(x)
}
\arguments{
\item{x}{An object (ExpressionSet, SummarizedExperiment, etc)}
}
\description{
Annotate an object with gene identifiers that are ensembl
}
\note{
This function currently only handles a SummarizedExperiment or child object.
}
