% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{methods}
\name{as_DFrame}
\alias{as_DFrame}
\alias{as_DFrame,tbl_df-method}
\alias{as_DFrame,tbl_df,ANY-method}
\title{Convert input to S4Vectors DataFrame}
\usage{
as_DFrame(x)

\S4method{as_DFrame}{tbl_df}(x)
}
\arguments{
\item{x}{The input to convert to DFrame}
}
\value{
A \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame()}} object.
}
\description{
Convert the input to a S4Vectors DataFrame object.
}
