% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.DFrame}
\alias{as_tibble.DFrame}
\title{Convert a DFrame to tibble}
\usage{
as_tibble.DFrame(x)
}
\arguments{
\item{x}{A \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DFrame()}} object}
}
\value{
A tibble representing \code{x}
}
\description{
Convert a DFrame to tibble
}
\details{
There are a few differences between a tibble and DFrame, some
of which are more involved. This function does not (yet) handle the
complex cases.

rownames: tibbles don't like rownames (or dplyr), so we store the rownames
in a variable called \code{.rownames} in the data frame. Should there be
a need, see \code{\link[tibble:rownames]{tibble::column_to_rownames()}} or \code{\link[=as_DFrame.tibble]{as_DFrame.tibble()}} for
options of getting the rownames back.
}
