% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_tibble.GRanges}
\alias{as_tibble.GRanges}
\title{Convert GRanges object to tibble}
\usage{
\method{as_tibble}{GRanges}(x)
}
\arguments{
\item{x}{A \code{\link[GenomicRanges:GRanges-class]{GenomicRanges::GRanges()}} object}
}
\value{
A tibble representing \code{x}
}
\description{
Convert GRanges object to tibble
}
