% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{methods}
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble,DFrame-method}
\alias{as_tibble,DFrame,ANY-method}
\alias{as_tibble,GRanges-method}
\alias{as_tibble,GRanges,ANY-method}
\title{Convert object to tibble}
\usage{
as_tibble(x)

\S4method{as_tibble}{DFrame}(x)

\S4method{as_tibble}{GRanges}(x)
}
\description{
Although the tibble class uses S3 for converting objects
to tibbles, the Bioconductor objects are S4. This converts tibbles that
are S4-oriented.
}
