% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{build_rsem_sample_table}
\alias{build_rsem_sample_table}
\title{Generate a RSEM-based sample table based on a directory}
\usage{
build_rsem_sample_table(dir, which = c("gene", "transcript"))
}
\arguments{
\item{dir}{The directory to look in for RSEM files.}

\item{which}{What RSEM output type (genes, isoforms) to identify}
}
\value{
A data.frame consisting of two columns \code{files} and \code{names}.
}
\description{
This function infers a set of samples that have RSEM output based on
a target directory (\code{dir}) and a RSEM type (\code{genes} or \code{isoforms}). The
sample name is then inferred from these files, resulting in a data frame
with \code{files} and \code{names} columns.
}
