% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{methods}
\name{col_data}
\alias{col_data}
\alias{col_data,SummarizedExperiment-method}
\alias{col_data,SummarizedExperiment,ANY-method}
\title{Extract column metadata as a tibble}
\usage{
col_data(x)

\S4method{col_data}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{The object to extract column metadata from}
}
\value{
A tibble containing column metadata from \code{x}.
}
\description{
Extract the column metadata from an object and return
the result as a tibble.
}
