% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{methods}
\name{col_data<-}
\alias{col_data<-}
\alias{col_data<-,SummarizedExperiment-method}
\alias{`col_data<-`,SummarizedExperiment,ANY-method}
\title{Set column metadata as a tibble}
\usage{
col_data(x) <- value

\S4method{col_data}{SummarizedExperiment}(x) <- value
}
\arguments{
\item{x}{The tibble of metadata to store.}
}
\value{
The object with column metadata included.
}
\description{
Some objects support column-oriented metadata. This function
will store this metadata with the object.
}
