% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{find_rsem_files}
\alias{find_rsem_files}
\title{Find RSEM files to load}
\usage{
find_rsem_files(dir)
}
\arguments{
\item{dir}{A directory to look in}
}
\description{
Identify files ending with \code{.genes.results} as RSEM files suitable for loading (with full
pathname).
}
