% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{gtf_gencode}
\alias{gtf_gencode}
\title{gtf}
\description{
Routines for interacting with gtf gene annotation files
}
\details{
GTF is a gene annotation format accessible from EBI/Ensembl. These
routines are wrappers around the excellent \link{rtracklayer} package. The goal
is merely to make working with these files a bit easier, in the mode of
the \link{usethis} package.

The GTF files (for gencode) have information at the gene, transcript and
exon levels. It is possible to get an object with all of these, or with
only data at one of these levels. These functions make the process slightly
easier.

\emph{Caching}: Since the files in question are large, one of the things that
this package does is provide caching (via BiocFileCache) of the files for
subsequent loads. Hopefully, it is often the case that the same file would
be needed multiple times.

Finally, although not very well thought out, there is a function (import_gencode_gtf)
which attempts to impose some logic on the process of inferring the right gencode
file. It is very much a work in progress.
}
