% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{gtf_gene_annotation}
\alias{gtf_gene_annotation}
\title{Extract gene-level annotation from gtf}
\usage{
gtf_gene_annotation(url, ...)
}
\arguments{
\item{gtf}{}
}
\description{
Extract gene-level annotation from gtf
}
