% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{gtf_transcript_annotation}
\alias{gtf_transcript_annotation}
\title{Extract transcript-level annotation from gtf}
\usage{
gtf_transcript_annotation(url, ...)
}
\arguments{
\item{gtf}{}
}
\description{
Extract transcript-level annotation from gtf
}
