% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{gtf_tx_annotation}
\alias{gtf_tx_annotation}
\title{Extract transcript-level annotation from gtf}
\usage{
gtf_tx_annotation(gtf)
}
\arguments{
\item{gtf}{}
}
\description{
Extract transcript-level annotation from gtf
}
