% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode.R
\name{import_gencode_gtf}
\alias{import_gencode_gtf}
\title{Import a gencode gtf}
\usage{
import_gencode_gtf(
  url = NULL,
  version = "v32",
  species = c("human"),
  gencode_base_url = "ftp://ftp.ebi.ac.uk/pub/databases/gencode"
)
}
\arguments{
\item{url}{A URL to use for direct file access. Note all other parameters
are ignored if this is set.}

\item{version}{A version number (assuming human)}

\item{species}{The species (currently human) to import}

\item{gencode_base_url}{The base URL of the genocode download (EBI)}
}
\value{
A GRanges object representing the genocode gtf (see \code{\link[rtracklayer:import]{rtracklayer::import()}}).
}
\description{
Import a gencode gtf
}
