% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem}
\alias{import_rsem}
\title{Import RSEM data using a sample table}
\usage{
import_rsem(sample_table, which = c("tx2gene", "gene", "tx"), tx2gene = NULL)
}
\arguments{
\item{tx2gene}{}
}
\description{
Import data from files using a sample table (with files and optionally names),
either at the transcript or gene level.
}
