% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem_as_DESeqDataSet}
\alias{import_rsem_as_DESeqDataSet}
\title{Import RSEM as SummarizedExperiment}
\usage{
import_rsem_as_DESeqDataSet(
  sample_table = ".",
  txIn = FALSE,
  txOut = FALSE,
  formula = ~1
)
}
\arguments{
\item{sample_table}{A data.frame of annotations and filenames. Can also be a path (default:".")
or a vector of filenames. If a data frame, should have one column (files) or have the variable
'files' for the filenames. There can also be a \code{names} variable which will be used for
sample names.}

\item{txIn}{logical - Transcript-level input (vs gene)}

\item{formula}{A formula for DESeq (default is ~1)}

\item{TxOut}{logical - Transcript-level output (vs gene)}
}
\value{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} representing the RSEM
data (with assays for counts and TPM).
}
\description{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} object (Counts, Normalized Counts and TPM).
}
