% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem_as_ExpressionSet}
\alias{import_rsem_as_ExpressionSet}
\title{Import RSEM as ExpressionSet}
\usage{
import_rsem_as_ExpressionSet(sample_table, formula = ~1, use_log = TRUE)
}
\arguments{
\item{sample_table}{A data.frame of annotations and filenames}

\item{formula}{A formula for DESeq (default is ~1)}

\item{use_log}{Should the data be log2 transformed?}
}
\value{
A \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} representing the RSEM data (with assayData for
exprs (normalized counts), counts (raw counts) and TPM).
}
\description{
\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} object (Counts, Normalized Counts and TPM).
}
