% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem_files}
\alias{import_rsem_files}
\title{Import RSEM data using tximport}
\usage{
import_rsem_files(
  files = ".",
  names = NULL,
  txIn = FALSE,
  txOut = FALSE,
  tx2gene_gtf = NULL,
  importer = NULL
)
}
\arguments{
\item{files}{A character vector of filenames to load. If not specified or if
the input is a directory name, all appropriate files are found in the specified
directory (or current directory).}

\item{names}{Optional character vector of sample names to use for files. If not
specified, the sample name from the filename will be inferred.}

\item{which}{Select either 'genes' or 'isoforms' for corresponding RSEM outputs}
}
\value{
A list (see \code{\link[tximport:tximport]{tximport::tximport()}}) of matrices containing RSEM data. Note
that the tx2gene option may be used, in which case a \code{tx2gene} data frame may
be included in the return result.
}
\description{
Import raw RSEM data into a list of quantifications.
}
\details{
This function takes a sample table containing the \code{files} field,
and uses \code{\link[tximport:tximport]{tximport::tximport()}} to load the files. The result is a list
containing several elements: abundance, counts, length. These are matrices
representing the RSEM data for a list of samples.
}
\note{
This function is focused on gene-level RSEM data, not transcript-level.
}
