% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{infer_rsem_samplename}
\alias{infer_rsem_samplename}
\title{Extract sample name from RSEM filename}
\usage{
infer_rsem_samplename(s)
}
\arguments{
\item{s}{A string (or vector of strings) representing filenames}
}
\value{
A vector (same length as s) consisting of sample names (stripped filenames).
}
\description{
Extract sample name from RSEM filename
}
