% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{infer_tx2gene_mapping}
\alias{infer_tx2gene_mapping}
\title{Infer tx2gene mapping for RSEM}
\usage{
infer_tx2gene_mapping(files = ".")
}
\arguments{
\item{files}{A list of RSEM transcript-level files}
}
\value{
A tx2gene mapping (two column data frame, first is tx, second is gene).
}
\description{
A simple approach of reading the tx2gene mapping directly from the
first RSEM input file provided. Not a great strategy but sometimes
this is all that is available.
}
