% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_estimate}
\alias{install_estimate}
\title{Install the ESTIMATE software (it is in a custom location)}
\usage{
install_estimate()
}
\value{
Invisible NULL (from install.packages).
}
\description{
Install the ESTIMATE software (it is in a custom location)
}
