% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_xcell}
\alias{install_xcell}
\title{Install the XCell software (it is in a custom location)}
\usage{
install_xcell()
}
\value{
Invisible NULL (from install.packages).
}
\description{
Install the XCell software (it is in a custom location)
}
