% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{normalize_rsem}
\alias{normalize_rsem}
\title{Title}
\usage{
normalize_rsem(rsem, sample_table, formula = ~1)
}
\arguments{
\item{rsem}{RSEM object (list with counts, etc)}

\item{sample_table}{Sample table (metadata) for samples.}

\item{formula}{Formaula for DESeq (default ~1)}
}
\description{
Title
}
