% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gct.R
\name{readGCT}
\alias{readGCT}
\title{Read a GCT-formatted file into an ExpressionSet object.}
\usage{
readGCT(file)
}
\arguments{
\item{file}{The name of the file (GCT format) to read in.}
}
\value{
An ExpressionSet representing the GCT data.
}
\description{
Read a GCT-formatted file into an ExpressionSet object.
}
\examples{
eset<-readGCT("/tmp/data.gct")

}
