% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{remove_empty_rows_from_rsem}
\alias{remove_empty_rows_from_rsem}
\title{Removes genes with no observations from RSEM object}
\usage{
remove_empty_rows_from_rsem(x, remove.lengths = TRUE, min.zero.obs = Inf)
}
\arguments{
\item{x}{An RSEM object}

\item{remove.lengths}{(TRUE) Remove empty lengths}

\item{min.zero.obs}{(Inf) Number of zero observations to trigger removing row}
}
\value{
An RSEM object with empty rows removed.
}
\description{
Removes rows (genes) that have length of 0 for any of the samples.
}
\details{
There appears to be a need in DESeq2 to ensure that no gene have an estimated
gene length of 0. This routine filters out those rows from an RSEM object.
}
