% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\docType{methods}
\name{row_data}
\alias{row_data}
\alias{row_data,SummarizedExperiment-method}
\alias{row_data,SummarizedExperiment,ANY-method}
\title{Extract row data from an object}
\usage{
row_data(x)

\S4method{row_data}{SummarizedExperiment}(x)
}
\arguments{
\item{x}{An object to extract row data from}
}
\value{
A tibble of row data
}
\description{
Extract row data from an object
}
