% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{tximport_rsem}
\alias{tximport_rsem}
\title{Import RSEM data using a sample table}
\usage{
tximport_rsem(
  sample_table,
  which = c("gene", "transcript", "tx2gene"),
  tx2gene = NULL
)
}
\arguments{
\item{tx2gene}{}
}
\description{
Import data from files using a sample table (with files and optionally names),
either at the transcript or gene level.
}
