% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{tximport_to_DESeq2}
\alias{tximport_to_DESeq2}
\title{Import tximport list to DESeq2 object}
\usage{
tximport_to_DESeq2(x, gene_annotation = NULL)
}
\arguments{
\item{x}{The tximport list from \code{\link[tximport:tximport]{tximport::tximport()}}}
}
\description{
Convert a tximport list of elements from \code{\link[tximport:tximport]{tximport::tximport()}} into a
DESeq2 object \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSet()}} for downstream processing.
}
