% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{tximport_to_ExpressionSet}
\alias{tximport_to_ExpressionSet}
\title{Title}
\usage{
tximport_to_ExpressionSet(x, phenoData = NULL, featureData = NULL)
}
\arguments{
\item{featureData}{}
}
\description{
Title
}
