% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{tximport_to_SummarizedExperiment}
\alias{tximport_to_SummarizedExperiment}
\title{Title}
\usage{
tximport_to_SummarizedExperiment(x, col_data = NULL, row_data = NULL)
}
\arguments{
\item{row_data}{}
}
\description{
Title
}
