% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{valid_tximport}
\alias{valid_tximport}
\title{Verify that tximport list is valid}
\usage{
valid_tximport(x)
}
\arguments{
\item{x}{A list from the \code{\link[tximport:tximport]{tximport::tximport()}} function}
}
\value{
A logical indicating of the tximport is valid.
}
\description{
Verify that tximport list is valid
}
