% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gct.R
\name{writeGCT}
\alias{writeGCT}
\title{Write an ExpressionSet to a GCT-formatted output file.}
\usage{
writeGCT(eset, file)
}
\arguments{
\item{eset}{The ExpressionSet to write.}

\item{file}{The file to write to.}
}
\description{
Write an ExpressionSet to a GCT-formatted output file.
}
\examples{
\dontrun{
writeGCT(eset, file="/tmp/exprs.gct")
}

}
