% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tibble.R
\name{as_DFrame.tibble}
\alias{as_DFrame.tibble}
\alias{as_DFrame.data.frame}
\title{Convert a tibble to a DFrame}
\usage{
\method{as_DFrame}{tibble}(x)

\method{as_DFrame}{data.frame}(x)
}
\arguments{
\item{x}{A tibble (or data.frame)}
}
\value{
A \code{\link[S4Vectors:DataFrame-class]{S4Vectors::DataFrame()}} object coerced from \code{x}.
}
\description{
Convert a tibble to a DFrame
}
\details{
The tibble and DFrame are similar, but there are important
differences. This function does not handle all of them (yet). The
biggest thing it handles is managing the conversion of rownames,
which can be non-unique in DFrame but not in tibble/data.frame. It
is assumed (if using \code{\link[=as_tibble.DFrame]{as_tibble.DFrame()}}) that the rownames are stored
in the data frame calumn \code{.rownames}. If it is not present, rownames are
not set. If non-unique, DFrame can handle this case by default.
}
\section{Functions}{
\itemize{
\item \code{as_DFrame(data.frame)}: Use a data frame

}}
