% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{find_rsem_files}
\alias{find_rsem_files}
\title{Find RSEM files to load}
\usage{
find_rsem_files(dir, which = c("gene", "transcript"))
}
\arguments{
\item{dir}{A directory to look in}

\item{which}{Select either 'genes' or 'isoforms' for corresponding RSEM outputs}
}
\value{
A character vector of full paths to RSEM output files
}
\description{
Find RSEM files ending with appropriate extensions for loading.
}
\details{
Identify files ending with \code{.genes.results} or \code{.isoforms.results} as RSEM
files suitable for loading (with full
pathname).
}
