% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem}
\alias{import_rsem}
\title{Title}
\usage{
import_rsem(
  sample_table,
  which = c("gene", "transcript", "tx2gene"),
  tx2gene = NULL,
  gene_annotation = NULL,
  importer = NULL
)
}
\description{
Title
}
