% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem}
\alias{import_rsem}
\title{Import RSEM data using tximport}
\usage{
import_rsem(files)
}
\arguments{
\item{files}{A character vector of filenames to load.}
}
\value{
A list (see \code{\link[tximport:tximport]{tximport::tximport()}}) of matrices containing RSEM data.
}
\description{
Import raw RSEM data into a list of quantifications.
}
\details{
This function takes a sample table containing the \code{filename} field,
and uses \code{\link[tximport:tximport]{tximport::tximport()}} to load the files. The result is a list
containing several elements: abundance, counts, length. These are matrices
representing the RSEM data for a list of samples.
}
\note{
This function is focused on gene-level RSEM data, not transcript-level.
}
\examples{
\dontrun{
import_rsem(c("foo.txt","bar.txt"))
}
}
