% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsem.R
\name{import_rsem_as_DESeqDataSet}
\alias{import_rsem_as_DESeqDataSet}
\title{Title}
\usage{
import_rsem_as_DESeqDataSet(sample_table, formula = ~1)
}
\arguments{
\item{sample_table}{A data.frame of annotations and filenames}

\item{formula}{A formula for DESeq (default is ~1)}
}
\value{
A \link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} representing the RSEM
data (with assays for counts and TPM).
}
\description{
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment} object (Counts, Normalized Counts and TPM).
}
